pyproject.toml
[project]
name = "htbrowser"
version = "0.1.0"
description = "A browser for the HathiTrust catalog"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.33.1",
]
